Autoencoder clustering

This function Compress data using autoencoder partially connected

autoencoderClustering(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  separator,
  nCluster,
  projectName,
  clusterMethod = c("GRIPH", "SIMLR", "SEURAT", "SHARP"),
  seed = 1111,
  pcaDimensions,
  permAtTime = 4,
  largeScale = FALSE
)

Arguments

group,

a character string. Two options: sudo or docker, depending to which group the user belongs

scratch.folder,

a character string indicating the path of the scratch folder

file,

a character string indicating the path of the file, with file name and extension included.Has to be the one in the projectName folder.Different so from the previous one.

separator,

separator used in count file, e.g. '\t', ','

nCluster,

number of cluster in which the dataset is divided

projectName,

might be different from the matrixname in order to perform different analysis on the same dataset

clusterMethod,

clustering methods: "GRIPH","SIMLR","SEURAT","SHARP"

seed,

important value to reproduce the same results with same input

pcaDimensions,

number of dimensions to use for Seurat Pca reduction.

permAtTime,

number of permutation in parallel

largeScale,

boolean for SIMLR analysis, TRUE if rows are less then columns or if the computational time are huge

Value

Author

Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino

Examples

if (FALSE) {
 autoencoderClustering(group="docker", scratch.folder="/home/user/Riccardo/Riccardo/1_inDocker_2/scratch", file="/home/user/Riccardo/Riccardo/1_inDocker_2/data/Results/testDocker/setA.csv",separator=",", nCluster=5,clusterMethod=c("SEURAT"),seed=1111,projectName="testDocker",13, largeScale = FALSE)
}