A function to extract priority lists of genes from ANOVAlike output

This function extracts clusters specific gene lists from ANOVAlike output. IMPORTANT: This analysis works only on up modulated genes, i.e. genes more expressed in the clusters under analysis with respect to the reference cluster

clustersFeatures(
  group = c("sudo", "docker"),
  fileLogFC,
  fileCounts,
  delta = 0.5,
  sep
)

Arguments

group,

a character string. Two options: sudo or docker, depending to which group the user belongs

fileLogFC,

a character string indicating the absolute path to logFC_filtered_DE_ file generated by anovaLike function

fileCounts,

a character string indicating the absolute path to reordered counts table file generated by anovaLike function

delta,

the minimal distance between the max value of FC for a gene in a cluster of interest and the nearest other max FC in any of the other clusters. This value define the minimal distance with respect to the same gene in an other cluster to identify it as a cluster specific gene.

sep,

separator used in count file, e.g. '\t', ','

Value

Three tab delimited files with prefix onlyUP\_, followed by the counts table name, i.e. the count table only containing the selected genes. onlyUP\_, followed by logFC_filtered_DE_, the table containing logFC only for the selected genes, onlyUP_clusters_specific_genes.txt, which contains the list of specific genes associated with the corresponding cluster.

Author

Raffaele Calogero, raffaele.calogero [at] unito [dot] it, University of Torino

Examples

if (FALSE) {
system("wget http://130.192.119.59/public/clusters.features.zip")
unzip("clusters.features.zip")
setwd("clusters.features")

    clustersFeatures(group="docker", fileLogFC=paste(getwd(),
       "logFC_filtered_DE_annotated_setPace_10000_noC5_reordered.txt",sep="/"),
        fileCounts=paste(getwd(),"annotated_setPace_10000_noC5_reordered.txt",sep="/"), delta=0.5, sep="\t")
}