A function allowing the identification of differentially expressed genes.

This function executes in a docker edgeR for the idnetification of differentially expressed genes in single-cells RNAseq

deDetection(
  group = c("sudo", "docker"),
  data.folder,
  counts.table,
  file.type = c("txt", "csv"),
  logFC.threshold = 1,
  FDR.threshold,
  logCPM.threshold = 4,
  plot = c(TRUE, FALSE)
)

Arguments

group,

a character string. Two options: sudo or docker, depending to which group the user belongs

data.folder,

a character string indicating the folder where input data are located and where output will be written

counts.table,

a character string indicating the counts table file. IMPORTANT in the header of the file the covariate group MUST be associated to the column name using underscore, e.g. cell1_cov1

file.type,

type of file: txt tab separated columns csv comma separated columns

logFC.threshold,

minimal logFC present in at least one of the comparisons with respect to reference covariate

FDR.threshold,

minimal FDR present in at least one of the comparisons with respect to reference covariate

logCPM.threshold,

minimal average abundance

plot,

TRUE if differentially expressed genes are represented in a plot.

Author

Raffaele Calogero, raffaele.calogero [at] unito [dot] it, University of Torino, Italy

Examples

if (FALSE) {
    #running deDetection
    system("wget http://130.192.119.59/public/buettner_counts_noSymb.txt.zip")
    unzip("buettner_counts_noSymb.txt.zip")
    lorenzFilter(group="docker", scratch.folder="/data/scratch/",
                data.folder=getwd(), matrixName="buettner_counts_noSymb",
                p_value=0.05, format="txt", separator='\t')

    system("wget ftp://ftp.ensembl.org/pub/release-92/gtf/mus_musculus/Mus_musculus.GRCm38.92.gtf.gz")
    system("gzip -d Mus_musculus.GRCm38.92.gtf.gz")
    scannobyGtf(group="docker", data.folder=getwd(),
                 counts.table="lorenz_buettner_counts_noSymb.txt",
                 gtf.name="Mus_musculus.GRCm38.92.gtf",
                 biotype="protein_coding", mt=FALSE, ribo.proteins=FALSE,
                 file.type="txt", umiXgene=3)

    deDetection(group="docker", data.folder=getwd(),
               counts.table="annotated_lorenz_buettner_counts_noSymb.txt",
               file.type="txt", logFC.threshold=1, FDR.threshold=0.05, logCPM.threshold=4, plot=TRUE)
}