Cells count table size

This function executes a ubuntu docker that counts row and colums of a counts table.

dimensions(group = c("sudo", "docker"), scratch.folder, file, separator)

Arguments

group,

a character string. Two options: sudo or docker, depending to which group the user belongs

scratch.folder,

a character string indicating the path of the scratch folder

file,

a character string indicating the path of the file, with file name and extension included

separator,

separator used in count file, e.g. '\t', ','

Value

a file called dimensions.txt containing the number of row and columns of a cells counts table

Author

Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino

Examples

if (FALSE) {
system("wget http://130.192.119.59/public/annotated_setPace_10000_noC5.txt.zip")
unzip("annotated_setPace_10000_noC5.txt.zip")
dimensions(group="docker", scratch.folder="/data/scratch",
            file=paste(getwd(),"annotated_setPace_10000_noC5.txt", sep="/"), separator="\t")
}