gibbsR

This function creates performs motif search in peptides. It was developed to use gibbscluster on TCR CDR3 peptide sequences

gibbsR(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  resFolderCustom = "NULL",
  jobName,
  nCluster,
  motifLength,
  maxDelLength,
  maxInsLength,
  numbOfSeed,
  penalityFactorIntCluster = 0.8,
  backGroundAminoFreq = 2,
  seqWeightType = 1
)

Arguments

group,
a character string. Two options: sudo or docker, depending to which group the user belongs

scratch.folder,
a character string indicating the path of the scratch folder

file,
path to peptide file,

resFolderCustom,
optional parameter. Default will store the results in fastqPath otherwise will store the results in resFolderCustom path.

jobName,
identify of the sample

nCluster,
number of clusters to be generated

motifLength,
length of the motif to be searched

maxDelLength,
max length of deletions in the peptide search, expressed in number of aa

maxInsLength,
max length of insertion in the peptide search, expressed in number of aa

numbOfSeed,
number of initial configurations to be used to detect the highest KLD sum

penalityFactorIntCluster,
default 0.8,

backGroundAminoFreq,
two value 1 defined by uniprot and 2 defined by the dataset under analysis

seqWeightType,
sequence weighting type: default 1, faster but less precise, 2 slower but more precise, 3 none

Value

an indexed genome compliant with 10XGenomics cellranger

Author

Luca Alessandrì

Examples

if (FALSE) {
library(rCASC)
dir.create("scratch")
scratch.folder=paste(getwd(),"scratch",sep="/")
fastqPath=paste(getwd(),"fastq",sep="/")
resFolder=paste(getwd(),"resFolder",sep="/")
dir.create(resFolder)

gibbsR(group="docker",scratch.folder=scratch.folder,file=file,resFolderCustom=results,jobName="test",nCluster=8,motifLength=4,maxDelLength=4,maxInsLength=4,numbOfSeed=4,penalityFactorIntCluster=0.8,backGroundAminoFreq=2,seqWeightType=1) 

newFolder=paste(results,"res",list.files(paste(results,"res",sep="/")),"cores",sep="/")
}