Imputing dropouts

This function executes a wrapper for scImpute: Accurate And Robust Inputation For Single Cell RNA-Seq Data (Li WV Nature Communications, vol. 9, Article number: 997, 2018)

impute(
  group = c("sudo", "docker"),
  data.folder,
  counts.matrix,
  drop.thre,
  cores,
  refining = FALSE
)

Arguments

group,
a character string. Two options: sudo or docker, depending to which group the user belongs

data.folder,
a character string indicating the folder where tab separated file of cells counts is located

counts.matrix,
a character string indicating the tab separated file of cells counts

drop.thre,
A number between 0 and 1, specifying the threshold to determine dropout values

cores,
a integer specifying the number of cores used for parallel computation.

refining,
boolean FALSE, TRUE, if TRUE it execute again scImpute but setting a different drop.thre. Rerunning to change drop.thre hase to be done in the same folder where the first run was done.

Value

A matrix file with imputed data.

Examples

if (FALSE) {
    #downloading fastq files
    system("wget http://130.192.119.59/public/singlecells_counts.txt.gz")
    system("gzip -d singlecells_counts.txt.gz")
    cascImpute(group="docker", data.folder=getwd(),
         counts.matrix="singlecells_counts.txt",
         drop.thre=0.5, cores=8, refining=FALSE)
}