Evaluation of the numbur of optimal clusters

This function analyze the data that came up from permutationClustering script.

nClusterEvaluationSIMLR(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  range1,
  range2,
  separator,
  logTen,
  nPerm,
  percent,
  rangeFile
)

Arguments

group,

a character string. Two options: sudo or docker, depending to which group the user belongs

scratch.folder,

a character string indicating the path of the scratch folder

file,

a character string indicating the path of the file, with file name and extension included

range1,

First number of cluster that has to be analyzed

range2,

Last number of cluster that has to be analyzed

separator,

separator used in count file, e.g. '\t', ','

logTen,

0 for raw count, 1 otherwise

nPerm,

how many times number of cluster will be evaluated

percent,

number of cells that has to be removed for bootstrap

rangeFile,

name of the file that has for each row each number of cluster that has to be analyzed

Value

stability plot for each nCluster,two files with score information for each cell for each permutation.

Author

Luca Alessandri , alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino

Examples

if (FALSE) {
permAnalysis("docker","path/to/scratch","path/to/data/TOTAL",3,4,",",0.8)#
}