Seurat Integration Permutation

This function executes a ubuntu docker that performs seurat integration to identify associated clusters in two independent experiment. The analysis is repeated mutiple times removing 10

seuratIntegrationPermutation(
  group = c("sudo", "docker"),
  scratch.folder,
  file1,
  file2,
  separator1,
  separator2,
  cl1,
  cl2,
  permutation,
  seed,
  outputFolder,
  K = 0.8
)

Arguments

group,

a character string. Two options: sudo or docker, depending to which group the user belongs

scratch.folder,

a character string indicating the path of the scratch folder

file1,

a character string indicating the path of the first matrix

file2,

a character string indicating the path of the second matrix

separator1,

separator used in count file, e.g. '\t', ','

separator2,

separator used in count file, e.g. '\t', ','

cl1,

path of clustering.output for file1

cl2,

path of clustering.output for file2

permutation,

number of permutation for statistic

seed,

integer file necessary for reproducibility

outputFolder,

path to the output folder

K,

resolution for seurat analysis

Value

A folder called ISC, the input data, the intermediate results and a comma separated file final_score.csv summarising the frequency by which the seuratIntegration function return a an integration cluster including at least 50

Author

Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino

Examples

if (FALSE) {
seuratIntegrationPermutation(group="docker", scratch.folder="/home/user/scratch", file1="/home/user/dockerFile/Seurat_join_DAPUSHARE/function/example/set1.csv",file2="/home/user/dockerFile/Seurat_join_DAPUSHARE/function/example/setA.csv", separator1=",",separator2=",",cl1=, cl29,permutation=100, seed=111) 
}