Spatial transcriptomics pipeline

Create count matrix from spatial transcriptomics fasta

stpipeline(
  group = c("sudo", "docker"),
  scratch.folder,
  data.folder,
  genome.folder,
  fastqPathFolder,
  ID,
  imgNameAndPath,
  slide = NULL,
  area = NULL
)

Arguments

group,

a character string. Two options: sudo or docker, depending to which group the user belongs

scratch.folder,

a character string indicating the path of the scratch folder

data.folder,

a character string indicating the path of the result folder.

genome.folder,

a character string indicating the path of the genome folder.

fastqPathFolder,

a character string indicating the path of fastq folder

ID,

a character string indicating the name of the project

imgNameAndPath,

path and name of tiff image file required for analysis.

slide,

identificative number from dataset download

area,

identificative value from dataset download

Value

count matrix from spatial transcriptomics

Author

Luca Alessandri , alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino

Examples

if (FALSE) {
Dataset="curl -O http://s3-us-west-2.amazonaws.com/10x.files/samples/spatial-exp/1.0.0/V1_Mouse_Kidney/V1_Mouse_Kidney_fastqs.tar"
DatasetImage="curl -O http://cf.10xgenomics.com/samples/spatial-exp/1.0.0/V1_Mouse_Kidney/V1_Mouse_Kidney_image.tif"
referenceGenomeHG38="curl -O http://cf.10xgenomics.com/supp/spatial-exp/refdata-cellranger-GRCh38-3.0.0.tar.gz"
referenceGenomeMM10="curl -O http://cf.10xgenomics.com/supp/spatial-exp/refdata-cellranger-mm10-3.0.0.tar.gz"
stpipeline(group="docker", scratch.folder="/run/media/user/Maxtor4/scratch", data.folder="/run/media/user/Maxtor4/prova2", genome.folder="/home/user/spatial/refdata-cellranger-mm10-3.0.0", fastqPathFolder="/home/user/spatial/V1_Mouse_Kidney_fastqs", ID="hey",imgNameAndPath="/home/user/spatial/V1_Mouse_Kidney_image.tif",slide="V19L29-096",area="B1")
}