integrationCircos

This function execute toprnk analysis which search for correspondence between clusters of two different experiments requires that the data are clustered with any of the software implemented in rCASC, cometsc, bulkClusters and autoencoder4pseudoBulk were already executed.

wrapperClustersIntegration(
  group = c("sudo", "docker"),
  scratch.folder,
  file.matrix1,
  file.matrix2,
  file.total1,
  file.total2,
  cl1,
  cl2,
  separator1,
  separator2,
  permutation = 100,
  seed = 111,
  top.ranked = 320,
  gsea = "msigdb.all",
  X = 5,
  L = 0.15,
  pvalue = 0.05,
  outputFolder
)

Arguments

group,

a character string. Two options: sudo or docker, depending to which group the user belongs

scratch.folder,

a character string indicating the path of the scratch folder

file.matrix1,

a character string indicating the path of the first matrix

file.matrix2,

a character string indicating the path of the second matrix

file.total1,

a character string indicating the path to the total.csv.for the 1st dataset to be integrated. Total.csv is generated with autoencoder4pseudoBulk. File, with file name and extension included.

file.total2,

a character string indicating the path to the total.csv.for the 2nd dataset to be integrated. Total.csv is generated with autoencoder4pseudoBulk. File, with file name and extension included.

cl1,

path of clustering.output for file.matrix1

cl2,

path of clustering.output for file.matrix2

separator1,

separator used in count file, e.g. '\t', ','

separator2,

separator used in count file, e.g. '\t', ','

permutation,

number of permutation to be run

seed,

integer file necessary for reproducibility

top.ranked,

MAX number of top comet genes to be used for each cluster, default 320

gsea,

default msigdb.all, which includes all classes. List of the available GSEA classes: c1.all, c2.cgp, c2.cp.biocarta, c2.cp.kegg, c2.cp.pid, c2.cp.reactome, c2.cp.wikipathways, c3.all, c3.mir, c3.tft.gtrd, c3.tft, c4.cgn, c4.cm, c5.go.bp, c5.go.cc, c5.go.mf, c5.hpo, c6.all, c7.all, c8.all, h.all, msigdb.all. Please note that msigdb.all includes all gsea classes.

X,

X parameter for the XLmHG, default 5, for more info please see XLmHG help: https://xl-mhg.readthedocs.io/en/latest/.

L,

L parameter for the XLmHG, default 0.15, for more info please see XLmHG help: https://xl-mhg.readthedocs.io/en/latest/.

pvalue,

XLmHG pvalue threshold,default 0.05

outputFolder,

where results are placed

Value

A picture called integrated_score.png and a file called integrated_score.csv and all the final_scores.csv used to produce the integrated results.

Author

Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino

Examples

if (FALSE) {
 library(rCASC)
 wrapperClustersIntegration(group="docker", 
        scratch.folder="/scratch", 
        file.matrix1="/data/clusters_association_paper/setA1_set1/setA1/VandE/VandE.csv",
        file.matrix2="/data/clusters_association_paper/setA1_set1/set1/VandE/VandE.csv",
        cl1="/data/clusters_association_paper/setA1_set1/setA1/VandE/Results/VandE/5/VandE_clustering.output.csv",
        cl2="/data/clusters_association_paper/setA1_set1/set1/VandE/Results/VandE/4/VandE_clustering.output.csv",
        file.total1="/data/clusters_association_paper/setA1_set1/setA1/VandE/Results/setA1/permutation/total.csv",
        file.total2="/data/clusters_association_paper/setA1_set1/set1/VandE/Results/set1/permutation/total.csv",
        separator1=",", separator2=",", 
        permutation=100, seed=111, top.ranked=320, gsea="msigdb.all", X=5, L=0.15, pvalue=0.05,
        outputFolder="/data/clusters_association_paper/setA1_set1"
        )
}