Cellranger indexing

This function creates the indexing for 10Xgenomics

cellrangerIndexing(
  group = c("sudo", "docker"),
  scratch.folder,
  genomeFolder,
  gtf.url,
  fasta.url,
  bio.type = c("protein_coding", "unitary_pseudogene", "unprocessed_pseudogene",
    "processed_pseudogene", "transcribed_unprocessed_pseudogene", "processed_transcript",
    "antisense", "transcribed_unitary_pseudogene", "polymorphic_pseudogene", "lincRNA",
    "sense_intronic", "transcribed_processed_pseudogene", "sense_overlapping",
    "IG_V_pseudogene", "pseudogene", "TR_V_gene", "3prime_overlapping_ncRNA",
    "IG_V_gene", "bidirectional_promoter_lncRNA", "snRNA", "miRNA", "misc_RNA", "snoRNA",
    "rRNA", "IG_C_gene", "IG_J_gene", 
     "TR_J_gene", "TR_C_gene", "TR_V_pseudogene",
    "TR_J_pseudogene", "IG_D_gene", "ribozyme", "IG_C_pseudogene", "TR_D_gene", "TEC",
    "IG_J_pseudogene", "scRNA", "scaRNA", "vaultRNA", "sRNA", "macro_lncRNA",
    "non_coding", "IG_pseudogene"),
  nThreads,
  version = "5"
)

Arguments

group,

a character string. Two options: sudo or docker, depending to which group the user belongs

scratch.folder,

a character string indicating the path of the scratch folder

genomeFolder,

path for the genome folder

gtf.url,

a character string indicating the URL from ENSEMBL ftp for the GTF for genome of interest

fasta.url,

a character string indicating the URL from ENSEMBL ftp for the unmasked genome sequence of interest

bio.type,

ENSEMBL biotype to filter the GTF

nThreads,

number of cores for parallelization

version,

cellranger version: 2, 3 or 5.

Value

an indexed genome compliant with 10XGenomics cellranger

Author

Luca Alessandrì

Examples

if (FALSE) {
library(rCASC)
setwd("/data/genomes/hg38refcellranger")

cellrangerIndexing(group="docker", scratch.folder="/data/scratch", 
            gtf.url="ftp://ftp.ensembl.org/pub/release-87/gtf/homo_sapiens/Homo_sapiens.GRCh38.87.gtf.gz",
            fasta.url="ftp://ftp.ensembl.org/pub/release-87/fasta/homo_sapiens/dna/Homo_sapiens.GRCh38.dna.toplevel.fa.gz",
            genomeFolder = getwd(), bio.type="protein_coding", nThreads = 8)
}