gseaXLmHG

This function execute XLmHG estimation of the presence of enriched GSEA class in cluster-specific genes detected by COMET for the clusters of two independent datasets and uses this information to search for correspondence between clusters in the two independent datasets. This function requires that the two datasets have been clustered in rCASC and COMET analysis was run.

gseaXLmHG(
  group = c("sudo", "docker"),
  scratch.folder,
  xCometFolder,
  yCometFolder,
  gsea = c("c1.all", "c2.cgp", "c2.cp.biocarta", "c2.cp.kegg", "c2.cp.pid",
    "c2.cp.reactome", "c2.cp.wikipathways", "c3.all", "c3.mir", "c3.tft.gtrd", "c3.tft",
    "c4.cgn", "c4.cm", "c5.go.bp", "c5.go.cc", "c5.go.mf", "c5.hpo", "c6.all", "c7.all",
    "c8.all", "h.all", "msigdb.all"),
  X,
  L,
  pvalue,
  separatorX,
  separatorY,
  outputFolder
)

Arguments

group,
a character string. Two options: sudo or docker, depending to which group the user belongs

scratch.folder,
a character string indicating the path of the scratch folder

xCometFolder,
path of Comet results from X experiment

yCometFolder,
path of Comet results from Y experiment

gsea,
list of the available GSEA classes: c1.all, c2.cgp, c2.cp.biocarta, c2.cp.kegg, c2.cp.pid, c2.cp.reactome, c2.cp.wikipathways, c3.all, c3.mir, c3.tft.gtrd, c3.tft, c4.cgn, c4.cm, c5.go.bp, c5.go.cc, c5.go.mf, c5.hpo, c6.all, c7.all, c8.all, h.all, msigdb.all. Please note that msigdb.all includes all gsea classes, we suggest to use this dataset for a complete analysis.

X,
X parameter for the XLmHG, default 5, for more info please see XLmHG help: https://xl-mhg.readthedocs.io/en/latest/.

L,
L parameter for the XLmHG, default 0.15, for more info please see XLmHG help: https://xl-mhg.readthedocs.io/en/latest/.

pvalue,
XLmHG pvalue threshold

separatorX,
separator used in count file, e.g. '\t', ','

separatorY,
separator used in count file, e.g. '\t', ','

outputFolder,
the folder where GSEA folder will be created

Value

A folder called GSEA with all the results generated. The function produces an integrated output, final_score.csv, of the results obtained for progressively increasing length of the cluster-specific-genes detected by COMET.

Author

Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino

Examples

if (FALSE) {
 library(rCASC)
 gseaXLmHG(group="docker",
        scratch.folder="/scratch", 
        xCometFolder="/data/reanalysis_on_AIsc/comparing_CRC0327/NT_CTX/CRC0327_NT_2_clx/VandE/Results/VandE/8/outputdata",
        yCometFolder="/data/reanalysis_on_AIsc/comparing_CRC0327/NT_CTX/CRC0327_cetux_2_clx/VandE/Results/VandE/8/outputdata",
        gsea="c2.cp.kegg",
        X=5,
        L=0.15,
        pvalue=0.05,
        separatorX=",",
        separatorY=",",
        outputFolder="/data/reanalysis_on_AIsc/comparing_CRC0327/NT_CTX"
 )
}