A function to handle a indrop V2 single cell data

This function executes a docker that produces as output the sinngle cell counts from V2 indrop single cell sequencing

indropCounts(
  group = c("sudo", "docker"),
  scratch.folder,
  fastq.folder,
  index.folder,
  sample.name,
  split.affixes,
  bowtie.index.prefix = "genome",
  M = 10,
  U = 2,
  D = 400,
  low.complexity.mask = "False",
  umi.threshold = 3
)

Arguments

group,

a character string. Two options: sudo or docker, depending to which group the user belongs

scratch.folder,

a character string indicating the path of the scratch folder

fastq.folder,

a character string indicating the folder where input data are located and where output will be written

index.folder,

a character string indicating the folder where transcriptome index was created with indropIndex.

sample.name,

the string indicating the sample name

split.affixes,

the string separating SAMPLENAME from the Rz_001.fastq.gz

bowtie.index.prefix,

the prefix name of the bowtie index. If genome was generated with indropIndex function the bowtie index is genome (default).

M,

Ignore reads with more than M alignments, after filtering on distance from transcript end.

U,

Ignore counts from UMI that should be split among more than U genes.

D,

Maximal distance from transcript end, NOT INCLUDING THE POLYA TAIL.

low.complexity.mask,

low complexity mask False (default) or True

umi.threshold,

the minimal number of UMI to consider a gene present

Author

Raffaele Calogero and Riccardo Panero, raffaele.calogero [at] unito [dot] it, Bioinformatics and Genomics unit, University of Torino Italy

Examples

if (FALSE) {
system("wget 130.192.119.59/public/testMm_S0_L001_R1_001.fastq.gz")
system("wget 130.192.119.59/public/testMm_S0_L001_R2_001.fastq.gz")
library(rCASC)
#running indropCounts
indropCounts(group="docker", scratch.folder="/data/scratch", fastq.folder=getwd(),
       index.folder="/data/genomes/mm10indrop",sample.name="testMm",
       split.affixes="S0_L001", bowtie.index.prefix="genome",
       M=10, U=2, D=400, low.complexity.mask="False", umi.threshold=3)
}